Added review of users

+++ src/main/java/delivery/stork/model/entity/User.java
@@ -27,7 +27,7 @@ public class User {
     @NonNull
     String password;
 
-    int averageRate;
+    double averageRate;
     @NonNull
     String instagram;
     @CreationTimestamp
